requirements.txt
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.3.2
cron-descriptor==1.4.0
decorator==5.1.1
Django==4.2.7
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-cron==0.6.0
django-crontab==0.7.1
django-filter==23.5
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
drf-yasg==1.21.7
eventlet==0.33.3
executing==2.0.1
flake8==6.1.0
greenlet==3.0.1
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kombu==5.3.4
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
packaging==23.2
parso==0.8.3
Pillow==10.1.0
pluggy==1.3.0
prompt-toolkit==3.0.41
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.17.2
PyJWT==2.8.0
pyproject==1.3.1
pyproject-toml==0.0.10
pyTelegramBotAPI==4.14.1
pytest==7.4.3
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.1
requests==2.31.0
rpds-py==0.16.2
shell==1.0.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
stripe==7.7.0
toml==0.10.2
traitlets==5.14.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
telebot~=0.0.5